:fire: Remove VoteView
