+++ Sources/Present/Detail/Type/ClassifyVoteButtonType.swift
