+++ Sources/Present/Main/Home/Model/VoteModel.swift
